awsiot: correct irregular use of 'status' event
The status event was being used to get details of the changes to the state,
whereas its role is really to report on the result of operations on the state
(shadow). We can/should rely on just the 'delta' event for change (message)
details.

Other minor changes/fixes to go along with the above.